requirements.txt
librosa>=0.10.0          # Audio analysis & pitch/tempo shifting
pydub>=0.25.1            # Audio manipulation
soundfile>=0.12.1        # Audio file I/O
suno-api>=0.0.1          # Suno unofficial API (cookie-based)
google-auth-oauthlib>=1.0.0  # YouTube OAuth
google-auth-httplib2>=0.1.0  # HTTP for Google APIs
google-api-python-client>=2.0.0  # YouTube API
requests>=2.28.0         # HTTP calls
tenacity>=8.0.0          # Retry logic
python-dotenv>=1.0.0     # .env loading
ffmpeg-python>=0.2.0     # ffmpeg wrapper
pydantic>=2.0.0          # Data validation
numpy>=1.21.0            # Required for librosa
scipy>=1.7.0             # Required for librosa
selenium>=4.15.0         # Browser automation
webdriver-manager>=4.0.0 # WebDriver management
yt-dlp>=2023.11.0        # Audio/video downloading
